requirements.txt
fastapi
uvicorn
pydantic
httpx
pytest
google-generativeai
